FastAPI implementation/requirements.txt
fastapi==0.108.0
uvicorn==0.20.0
sqlalchemy
pydantic==1.10.4
python-multipart==0.0.5
aiofiles==22.1.0
python-dotenv==0.21.1
psycopg2-binary==2.9.6
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4
influxdb-client==1.36.0
aio-pika==8.3.0
pika==1.3.1
ansible==7.2.0
argon2_cffi==21.3.0
alembic
minio==7.1.13
fastapi-utils
opencv-python==4.8.1.78
packaging==23.2
pamqp==3.2.1
loguru
asyncpg
sqlalchemy-utils
fastapi-admin
websockets
requests
celery
redis
fastapi-cache2
fastapi-pagination

sentry-sdk[fastapi]
Pillow
geopy
gunicorn
pymongo
motor
httpx

pandas
scikit-learn
openpyxl
rarfile
XlsxWriter
pyfcm
